divided realisation and declaration of analysis
